+++ app/src/main/AndroidManifest.xml
@@ -17,9 +17,6 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.CaZaIt"
         tools:targetApi="31">
-        <activity
-            android:name=".ui.review.ReviewEditActivity"
-            android:exported="false" />
         <meta-data
             android:name="com.naver.maps.map.CLIENT_ID"
             android:value="@string/naver_client_id" />
