+++ app/src/main/res/navigation/nav_graph.xml
@@ -12,7 +12,7 @@
         tools:layout="@layout/fragment_cafe_list">
         <action
             android:id="@+id/action_cafeListFragment_to_cafeInfoFragment"
-            app:destination="@id/cafeInfoFragment"/>
+            app:destination="@id/cafeInfoFragment" />
         <action
             android:id="@+id/action_cafeListFragment_to_searchFragment"
             app:destination="@+id/searchFragment" />
@@ -24,7 +24,7 @@
         tools:layout="@layout/fragment_cafe_map">
         <action
             android:id="@+id/action_cafeMapFragment_to_cafeInfoFragment"
-            app:destination="@id/cafeInfoFragment"/>
+            app:destination="@id/cafeInfoFragment" />
     </fragment>
     <fragment
         android:id="@+id/myPageFragment"
@@ -33,7 +33,7 @@
         tools:layout="@layout/fragment_my_page">
         <action
             android:id="@+id/action_myPageFragment_to_recentlyCafeFragment"
-            app:destination="@id/recentlyCafeFragment"/>
+            app:destination="@id/recentlyCafeFragment" />
         <action
             android:id="@+id/action_myPageFragment_to_signInFragment"
             app:destination="@+id/signInFragment" />
@@ -165,6 +165,25 @@
         <argument
             android:name="cafe"
             app:argType="org.cazait.model.Cafe" />
+        <action
+            android:id="@+id/action_cafeInfoFragment_to_reviewWriteFragment"
+            app:destination="@+id/reviewWriteFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/reviewWriteFragment"
+        android:name="org.cazait.ui.review.ReviewWriteFragment"
+        android:label="ReviewWriteFragment"
+        tools:layout="@layout/fragment_review_write">
+        <argument
+            android:name="cafe"
+            app:argType="org.cazait.model.Cafe" />
+        <argument
+            android:name="score"
+            app:argType="float" />
+        <argument
+            android:name="reviewContent"
+            app:argType="string"
+            app:nullable="true" />
     </fragment>
     <fragment
         android:id="@+id/searchFragment"
